+++ src/main/java/net/silentchaos512/mechanisms/block/dryingrack/DryingRackTileEntity.java
@@ -11,17 +11,26 @@ import net.minecraft.network.play.server.SUpdateTileEntityPacket;
 import net.minecraft.particles.ParticleTypes;
 import net.minecraft.tileentity.ITickableTileEntity;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.Direction;
 import net.minecraft.util.NonNullList;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.util.LazyOptional;
+import net.minecraftforge.items.CapabilityItemHandler;
+import net.minecraftforge.items.IItemHandler;
+import net.minecraftforge.items.wrapper.InvWrapper;
 import net.silentchaos512.lib.util.PlayerUtils;
 import net.silentchaos512.mechanisms.crafting.recipe.DryingRecipe;
 import net.silentchaos512.mechanisms.init.ModTileEntities;
 import net.silentchaos512.mechanisms.util.ParticleUtils;
 
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 public class DryingRackTileEntity extends TileEntity implements IInventory, ITickableTileEntity {
     private final NonNullList<ItemStack> items = NonNullList.withSize(1, ItemStack.EMPTY);
     private int processTime;
+
+    private final LazyOptional<IItemHandler> itemHandlerCap = LazyOptional.of(() -> new InvWrapper(this));
 
     public DryingRackTileEntity() {
         super(ModTileEntities.dryingRack);
@@ -177,5 +186,19 @@ public class DryingRackTileEntity extends TileEntity implements IInventory, ITic
         if (pkt.getNbtCompound().contains("Item")) {
             setInventorySlotContents(0, ItemStack.read(pkt.getNbtCompound().getCompound("Item")));
         }
+    }
+
+    @Override
+    public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable Direction side) {
+        if (!this.removed && side != null && cap == CapabilityItemHandler.ITEM_HANDLER_CAPABILITY) {
+            return itemHandlerCap.cast();
+        }
+        return super.getCapability(cap, side);
+    }
+
+    @Override
+    public void remove() {
+        super.remove();
+        itemHandlerCap.invalidate();
     }
 }