+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Wires not connecting to some blocks (more work needed) [#62]
+- Drying rack not interacting with some item transfer methods (missing item handler capability) [#55]
 
 ## [0.6.12] - 2020-01-21
 ### Removed